# Optimize jumps in PartialOrd le