… src/test/codegen/issue-73338-effecient-cmp.rs