@AngelicosPhosphoros 2021-04-03
Probably need to run benchmarks suite for this to be sure that it doesn't break other optimizations.